pyproject.toml
[tool.poetry]
name = "ex-8-2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
mongoengine = "^0.27.0"
faker = "^22.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
